test: move pydantic related tests from test_build
